stock_monitoring/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
idna==3.7
packaging==24.0
PyJWT==2.8.0
requests==2.32.1
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
whitenoise==6.6.0
